Java_Painter
Cool painting tool, written in Java, based on the Processing library.
Sources are available for Eclipse, as well as for the Processing IDE.
The GUI Elements and the GUI Manager is also an own implementation (not
Swing, AWT or some Java standard GUI Lib) - also realized with the
Processing library.